requirements.txt
behave==1.2.5
behave-django==1.0.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-rest-framework==0.1.0
djangorestframework==3.7.7
idna==2.6
mysqlclient==1.3.12
parse==1.8.2
parse-type==0.4.2
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
